% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ArrArrHexPlot}
\alias{ArrArrHexPlot}
\title{Plot the values in two arrays against each other.}
\usage{
ArrArrHexPlot(arr.x, arr.y, bins = 100, log.trans = FALSE)
}
\arguments{
\item{arr.x, arr.y}{The two arrays. The \code{arrx} values will be along the \eqn{x} axis and the \code{arry} values along the \eqn{y} axis.

This is a \code{ggplot2} object and can be manipulated thus.}
}
\description{
Plot the values of two arrays of identical dimension against each other using
a hexagonal heatmap.
}

