% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brightness.R
\name{Brightness}
\alias{Brightness}
\alias{BrightnessTxtFolder}
\title{Calculate brightness from image series.}
\usage{
Brightness(mat3d, tau = NA, mst = NULL, filt = NULL, verbose = TRUE)

BrightnessTxtFolder(folder.path = ".", tau = NA, mst = NULL,
  filt = NULL, ext = "\\\\.tif$", mcc = parallel::detectCores(),
  verbose = TRUE)
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series. To perform this on a file
that has not yet been read in, set this argument to the path to that file
(a string).}

\item{tau}{If this is specified, bleaching correction is performed with
]code{\link{CorrectForBleaching}} which uses exponential filtering with
time constant \code{tau} (where the unit of time is the time between
frames).}

\item{mst}{Do you want to apply an intensity threshold prior to calculating
brightness (via \code{\link{MedStackThresh}})? If so, set your thresholding
\emph{method} here.}

\item{filt}{Do you want to median/smooth filter (with a radius of 1) the
brightness image using \code{\link{MedianFilterB}} or
\code{\link{SmoothFilterB}}? If \code{filt = "median"},
\code{\link{MedianFilterB}} is invoked with the option \code{na_count =
TRUE}.}

\item{verbose}{If mat3d is specified as a file name, print a message to tell
the user that that file is now being processed (useful for
\code{BrightnessFolder}, does not work with multiple cores) and to tell
when \code{MeanIntensityTxtFolder} is done.}

\item{folder.path}{The path (relative or absolute) to the folder you wish to
process.}

\item{ext}{the file extension of the images in the folder that you wish to
process (can be rooted in regular expression for extra-safety, as in the
default). You must wish to process all files with this extension; if there
are files that you don't want to process, take them out of the folder.}

\item{mcc}{The number of parallel cores to use for the processing.}
}
\value{
\code{Brightness} returns a matrix, the brightness image. The result
  of \code{BrightnessTxtFolder} is the text csv files written to disk (in the
  same folder as the input images).
}
\description{
Given a time stack of images, \code{Brightness} performs a calculation of the
molecular brightness for each pixel. \code{BrightnessTxtFolder} does this
calculation for an entire folder, writing the results as text files via
\code{\link{WriteImageTxt}}.
}
\examples{
library(EBImage)
img <- ReadImageData(system.file("extdata",
"low_oligomers.tif",
package = "nandb"))
display(normalize(img[, , 1]), method = "raster")
brightness <- Brightness(img, tau = 10, mst = "Huang", filt = "median")
KmerPlot(brightness, 1.16)
}

