% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightness.R
\name{brightness_time_series_folder}
\alias{brightness_time_series_folder}
\title{Brightness time-series calculations for every image in a folder.}
\usage{
brightness_time_series_folder(folder_path = ".", def, frames_per_set,
  tau = NULL, thresh = NULL, fail = NA, filt = NULL, s = 1,
  offset = 0, readout_noise = 0, gamma = 1, seed = NULL,
  parallel = FALSE)
}
\arguments{
\item{folder_path}{The path (relative or absolute) to the folder you wish to
process.}

\item{def}{A character. Which definition of brightness do you want to use,
\code{"B"} or \code{"epsilon"}?}

\item{frames_per_set}{The number of frames with which to calculate the
successive numbers.}

\item{tau}{The \eqn{tau} parameter for \emph{exponential filtering} detrending.
This must be a positive number. Set this to "auto" to use Nolan's algorithm
to automatically find a suitable value for this parameter (recommended).
For multi-channel images, it is possible to have a different \code{tau} for each
channel by specifying \code{tau} as a vector or list.}

\item{thresh}{The threshold or thresholding method (see
\code{\link[autothresholdr:mean_stack_thresh]{autothresholdr::mean_stack_thresh()}}) to use on the image prior to
detrending and brightness calculations.}

\item{fail}{If thresholding is done, to which value should pixels not
exceeding the threshold be set?}

\item{filt}{Do you want to smooth (\code{filt = 'mean'}) or median (\code{filt = 'median'}) filter the number image using \code{\link[=smooth_filter]{smooth_filter()}} or
\code{\link[=median_filter]{median_filter()}} respectively? If selected, these are invoked here with a
filter radius of 1 (with corners included, so each median is the median of
9 elements) and with the option \code{na_count = TRUE}. If you want to
smooth/median filter the number image in a different way, first calculate
the numbers without filtering (\code{filt = NULL}) using this function and then
perform your desired filtering routine on the result.}

\item{s}{A number. The \eqn{S}-factor of microscope acquisition.}

\item{offset}{Microscope acquisition parameters. See reference
Dalal et al.}

\item{readout_noise}{Microscope acquisition parameters. See reference
Dalal et al.}

\item{gamma}{Factor for correction of number \eqn{n} due to the illumination
profile. The default (\code{gamma = 1}) has no effect. Changing gamma will have
the effect of dividing the result by \code{gamma}, so the result with \code{gamma = 0.5} is two times the result with \code{gamma = 1}. For a Gaussian illumination
profile, use \code{gamma = 0.3536}; for a Gaussian-Lorentzian illumination
profile, use \code{gamma = 0.0760}.}

\item{seed}{Random numbers may be generated during the detrending process.
For reproducibility, you can set a seed for this random number generation
here.}

\item{parallel}{Would you like to use multiple cores to speed up this
function? If so, set the number of cores here, or to use all available
cores, use \code{parallel = TRUE}.}
}
\description{
Perform \code{\link[=brightness_time_series]{brightness_time_series()}} calculations on all tif images in a folder
and save the resulting number images to disk.
}
\examples{
setwd(tempdir())
img <- ijtiff::read_tif(system.file('extdata', '50.tif', package = 'nandb'))
ijtiff::write_tif(img, 'img1.tif')
ijtiff::write_tif(img, 'img2.tif')
brightness_time_series_folder(def = "e", tau = NA, thresh = "Huang",
                              frames_per_set = 20, parallel = 2)
suppressWarnings(file.remove(list.files()))  # cleanup
}
\seealso{
\code{\link[=brightness_time_series]{brightness_time_series()}}
}
