% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{CellMoved}
\alias{CellMoved}
\title{Which pixels have too many consecutive low values?}
\usage{
CellMoved(mat3d, thresh, method = "Huang")
}
\arguments{
\item{mat3d}{A 3d matrix representing the image series.}

\item{thresh}{The threshold below which values are considered \emph{low}.}

\item{method}{We want to decide how many consecutive \emph{low} values are
too many. This will be done by using
\code{\link[autothresholdr:auto_thresh]{autothresholdr::auto_thresh()}}, so here you set the method for
that function.}
}
\value{
A logical matrix (or vector or array) with \code{TRUE} for pixels
that have too many consecutive low values and \code{FALSE} for the others
(so \code{TRUE} for the \emph{bad} pixels).
}
\description{
If a cell moves into a pixel during the course of a multi-frame acquisition,
as the frames go on, the intensity at that pixel will go from low (absence of
cell) to high (presence of cell). We detect this movement in a given pixel by
detecting the series of consecutive low values. The 'low' values are those
failing to exceed a certain threshold (set by thresholding the whole image
via \code{\link[autothresholdr:auto_thresh]{autothresholdr::auto_thresh()}}). Of course this also works for
cells moving out of pixels.
}
\details{
This assumes that you've already thresholded out pixels where the cell
\emph{never} was (perhaps via \code{\link[autothresholdr:auto_thresh]{autothresholdr::auto_thresh()}}),
setting those to \code{NA}.
}
