% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Closest}
\alias{Closest}
\title{What's the closest value in a vector?}
\usage{
Closest(x, vec, index = FALSE)
}
\arguments{
\item{x}{A number.}

\item{vec}{A numeric vector.}

\item{index}{If set to \code{TRUE}, return the index (rather than the value) of
the closest element.}
}
\value{
A number.
}
\description{
Given a number \code{x} and a numeric vector \code{vec}, what's the closest value to
\code{x} in \code{vec}?
}
\examples{
Closest(pi, 0:10)
Closest(pi, 0:10, index = TRUE)

}
