% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{display}
\alias{display}
\title{Image display.}
\usage{
display(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[fields:image.plot]{fields::image.plot()}}. These arguments should
be named.}
}
\description{
This function is imported from the \code{detrendr} package. See the full
documentation at \code{\link[detrendr:display]{detrendr::display()}}.
}
