% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpSmooth}
\alias{ExpSmooth}
\title{Exponentially smooth a series of observations.}
\usage{
ExpSmooth(obs, tau)
}
\arguments{
\item{obs}{A numeric vector of observations (in order).}

\item{tau}{The time scale for the exponential smoothing (see Stroud 1999).}
}
\value{
The smoothed series, a numeric vector of the same length.
}
\description{
This function assumes that the observations are evenly spaced and separated
by 1 time unit (so choose your \code{tau} based on that).
}
\examples{
ExpSmooth(1:10, 1)
}

