% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpSmoothPillars}
\alias{ExpSmoothPillars}
\title{Exponentially smooth pillars of a 3-dimensional array}
\usage{
ExpSmoothPillars(mat3d, tau)
}
\arguments{
\item{tau}{The time scale for the exponential smoothing (see Stroud 1999).}
}
\value{
A 3-dimensional array where each pillar has been smoothed.
}
\description{
For a 3-dimensional array \code{mat3d}, pillar \code{i,j} is defined as
\code{mat3d[i, j, ]}. This function performs \link{ExpSmooth} on each pillar.
}
\examples{
m3 <- array(1:12, dim = c(2, 2, 3))
ExpSmoothPillars(m3, 7)
}

