% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{KmersFromImage}
\alias{KmersFromImage}
\alias{KmersFromImagesFolder}
\title{Calculate numbers of kmers based on an image time-series}
\usage{
KmersFromImage(img, monomer.med, tau = NA, verbose = TRUE)

KmersFromImagesFolder(folder.path, tau, monomer.med, out.name = "results",
  ext = "\\\\.tif$", verbose = TRUE)
}
\arguments{
\item{img}{A 3-dimensional array that one would might input to
\code{\link{Brightness}} \emph{or} the path to an image file on disk.}

\item{monomer.med}{The median brightness of a monomer. You must know what
this is to use this function correctly. This must be greater than 1 (this
is the "apparent brightness" of a monomer). If you're reading the Digman et
al. (2008) paper, this is \eqn{1 + \epsilon}.}

\item{tau}{The time constant for the exponential filtering (see
\code{\link{Brightness}}).}

\item{verbose}{Do you want to print the image name as you process it?}

\item{out.name}{The name of the results csv file.}

\item{ext}{the file extension of the images in the folder that you wish to
process (rooted in regular expression). You must wish to process all files
with this extension; if there are files that you don't want to process,
take them out of the folder.}
}
\value{
A named vector (named "1mers" "2mers" "3mers" and so on) with each
  element detailing the number of that kmer found, or for
  \code{KmersFromImagesFolder}, a csv file is written to disk detailing one
  of these vectors for each image.
}
\description{
Given an image (as a file path or an array), \code{KmersFromImage} does the
brightness calculation via \code{\link{Brightness}} and then counts the
numbers of each kmer via \code{\link{KmersFromBrightnesses}}.
\code{KmersFromImagesFolder} does this for an entire folder (directory) of
images and outputs a csv file of the results.
}

