% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MedianFilterB}
\alias{MedianFilterB}
\title{Image median filter with options for dealing with NAs}
\usage{
MedianFilterB(mat, size = 1L, na_rm = FALSE, na_count = FALSE)
}
\arguments{
\item{mat}{A matrix (representing an image).}

\item{size}{An integer; the median filter radius.}

\item{na_rm}{Should \code{NA}s be ignored?}

\item{na_count}{If this is TRUE, in each median calculation, if the majority
of arguments are \code{NA}s, \code{NA} is returned but if the \code{NA}s are
in the minority, they are ignored as in \code{median(x, na.rm = TRUE)}.}
}
\value{
A matrix (the median filtered image).
}
\description{
This is an alternative to \link[EBImage]{EBImage}'s
\code{\link[EBImage]{medianFilter}}. \code{MedianFilterB} has more options
for dealing with NA values. Whilst \code{\link[EBImage]{medianFilter}} can
either ignore NAs or set the output of any median calculation involving an
\code{NA} to \code{NA}, \code{MedianFilterB} can deal with \code{NA}s
depending on how many of them there are in a given median calculation.
}
\details{
The behavior at image boundaries is such as the source image has been padded
with pixels whose values equal the nearest border pixel value.
}
\examples{
m <- matrix(1:9, nrow = 3)
m[2:3, 2:3] <- NA
print(m)
MedianFilterB(m)
MedianFilterB(m, na_rm = TRUE)
MedianFilterB(m, na_count = TRUE)

}

