% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{NumberTimeSeries}
\alias{NumberTimeSeries}
\title{Create a number time-series.}
\usage{
NumberTimeSeries(mat3d, frames.per.set, tau = NA, mst = NULL,
  skip.consts = FALSE, filt = NULL, verbose = FALSE,
  mcc = parallel::detectCores())
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series. To perform this on a file
that has not yet been read in, set this argument to the path to that file
(a string).}

\item{frames.per.set}{The number of frames with which to calculate the
successive numbers.}

\item{tau}{If this is specified, bleaching correction is performed with
\code{\link{CorrectForBleaching}} which uses exponential filtering with
time constant \code{tau} (where the unit of time is the time between
frames). If this is set to \code{'auto'}, then the value of \code{tau} is
calculated automatically via \code{\link{BestTau}}.}

\item{mst}{Do you want to apply an intensity threshold prior to calculating
number (via \code{\link{MeanStackThresh}})? If so, set your thresholding
\emph{method} here.}

\item{skip.consts}{An image array with only one value (a 'constant array')
won't threshold properly. By default the function would give an error, but
by setting this parameter to \code{TRUE}, the array would instead be
skipped (the function will return the original array) and give a warning.}

\item{filt}{Do you want to smooth (\code{filt = 'smooth'}) or median
(\code{filt = 'median'}) filter the number image using
\code{\link{SmoothFilterB}} or \code{\link{MedianFilterB}} respectively? If
selected, these are invoked here with a filter radius of 1 and with the
option \code{na_count = TRUE}. If you want to smooth/median filter the
number image in a different way, first calculate the numbers without
filtering (\code{filt = NULL}) using this function and then perform your
desired filtering routine on the result.}

\item{verbose}{If mat3d is specified as a file name, print a message to tell
the user that that file is now being processed (useful for
\code{NumberTxtFolder}, does not work with multiple cores).}

\item{mcc}{The number of cores to use for the parallel processing.}
}
\value{
An array where the \eqn{i}th slice is the \eqn{i}th number image.
}
\description{
Given a stack of images \code{mat3d}, use the first \code{frames.per.set} of them to
create one number image, the next \code{frames.per.set} of them to create
the next number image and so on to get a time-series of number
images. If \code{tau} is specified, bleaching correction is performed via
\code{\link{CorrectForBleaching}}.
}
\details{
This may discard some images, for example if 175 frames are in the input and
\code{frames.per.set = 50}, then the last 25 are discarded. If bleaching
correction is selected, it is performed on the whole image stack before the
sectioning is done for calculation of numbers.
}
\examples{
library(EBImage)
img <- ReadImageData(system.file('extdata', '50.tif', package = 'nandb'))
display(normalize(img[, , 1]), method = 'raster')
bts <- NumberTimeSeries(img, 10, tau = 'auto', mst = 'tri', filt = 'median', mcc = 2)

}
\seealso{
\code{\link{Number}}.
}
