% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Stack2DTifs}
\alias{Stack2DTifs}
\title{Put individual tif files into one tif stack.}
\usage{
Stack2DTifs(file.names, out.name)
}
\arguments{
\item{file.names}{The names of the files to stack (in order).}

\item{out.name}{The name of the output .tif file.}
}
\description{
Say you have saved what you would like to be one 3D tif stack as a series of 2D tif files. This helps you stack them into one file.
}

