% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{SubSide}
\alias{SubSide}
\title{Subtract one side of a distribution from both sides.}
\usage{
SubSide(samp, midpoint, side = "left", both = TRUE, negs = FALSE)
}
\arguments{
\item{samp}{An empirical sample of the distribution.}

\item{midpoint}{To say what we mean by a "side" of the distribution, we need
a middle to say "the left is anything to the left of the middle and the
right is anything to the right". This \code{midpoint} says where the
"middle" is.}

\item{side}{Which side do you wish to subtract, left or right? It's good
enough to use the first letter.}

\item{both}{Subtract the selected side from both sides of the distribution (\code{both = TRUE}), or just the other side (\code{both = FALSE})?}

\item{negs}{Allow negative values in the probability density of the result?}
}
\value{
A data frame with two columns x and y giving the probability density of the output. The probability density at point x is y. Beware that this is not a real probability distribution in that it doesn't integrate to 1 and indeed it allows negative probability densities.
}
\description{
Given a (user-specified) midpoint of a distribution, subtract one side of it
from the other or both (the default, making one side zero) side(s), subtracting from the other side by mirroring across the midpoint. Beware that this function does not normalize the output such that it integrates to 1.
}
\examples{
sampl <- 0:10
plot(density(sampl))
subside3 <- SubSide(sampl, 4)
plot(subside3, type = "l")
plot(SubSide(sampl, 3, both = FALSE), type = "l")
}

