% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{text-image-io}
\alias{text-image-io}
\alias{write_txt_img}
\alias{read_txt_img}
\title{Read/write an image array to/from disk as text file(s).}
\usage{
write_txt_img(img, path, rds = FALSE)

read_txt_img(path)
}
\arguments{
\item{img}{An image, represented by a 2- or 3-dimensional array.}

\item{path}{The name of the input/output output file(s), \emph{without}
a file extension.}

\item{rds}{In addition to writing a text file, save the image as an RDS (a
single R object) file?}
}
\description{
If (as with brightness) you wish for the pixel values in an image to be
represented by real numbers that aren't necessarily integers, the tiff format
won't work. As a workaround we represent images (arrays) as tab-separated
\code{.txt} files on disk, where for image stacks (3-dimensional arrays), we
write one file for each slice, numbering it with the slice number.
}
\examples{
setwd(tempdir())
img <- read_tif(system.file('extdata', '50.tif', package = 'nandb'))
write_txt_img(img, 'temp')

img <- read_txt_img('temp_01.txt')
file.remove(list.files())  # cleanup
}
