% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{tiff-io}
\alias{tiff-io}
\alias{read_tif}
\alias{write_tif}
\title{TIFF I/O.}
\usage{
read_tif(path, n_ch = 1)

write_tif(img, file_name, na = "error", rds = FALSE)
}
\arguments{
\item{path}{The path to the image file on disk.}

\item{n_ch}{The number of channels in the image.}

\item{img}{An integer array.}

\item{file_name}{The name of the tif file (with or without the .tif) that you
wish to write.}

\item{na}{How do you want to treat \code{NA} values? R can only write integer
values (and hence not \code{NA}s) to tiff pixels. \code{na = 'saturate'} sets them to
saturated value. \code{na = 'zero'} sets them to zero, while \code{na = 'error'} will
give an error if the image contains \code{NA}s. You can also specify directly
here a natural number (must be between 0 and 2 ^ 32 - 1) to use in place of
\code{NA}s.}

\item{rds}{In addition to writing a TIFF file, do you want to save an RDS (a
single R object) file?}
}
\description{
These functions are imported from the \code{detrendr} package. See their full
documentation at \code{\link[detrendr:read_tif]{detrendr::read_tif()}} and \code{\link[detrendr:write_tif]{detrendr::write_tif()}}.
}
