% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brightness.R
\name{Brightness}
\alias{Brightness}
\title{Calculate brightness from image series.}
\usage{
Brightness(mat3d, tau = NA)
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series.}

\item{tau}{The time constant for the exponential filtering.}
}
\value{
A matrix, the brightness image.
}
\description{
Given a time stack of images, this function performs a calculation of the
molecular brightness for each pixel. If \code{tau} is specified, bleaching
correction is performed via \code{\link{CorrectForBleaching}}.
}

