% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightness.R
\name{brightness}
\alias{brightness}
\title{Calculate brightness from image series.}
\usage{
brightness(img, def, n_ch = 1, tau = NULL, thresh = NULL, fail = NA,
  filt = NULL, s = 1, offset = 0, readout_noise = 0, gamma = 1,
  seed = NULL, parallel = FALSE)
}
\arguments{
\item{img}{The image series to be detrended. This must be a 3-dimensional
array of non-negative integers, with the matrix defined as \code{img[, , 1]}
being the first frame of the image series.}

\item{def}{A character. Which definition of brightness do you want to use,
\code{"B"} or \code{"epsilon"}?}

\item{n_ch}{The number of channels in the image (default 1).}

\item{tau}{The \eqn{tau} parameter for \emph{exponential filtering} detrending.
This must be a positive number. Set this to "auto" to use Nolan's algorithm
to automatically find a suitable value for this parameter (recommended).}

\item{thresh}{The threshold or thresholding method (see
\code{\link[autothresholdr:mean_stack_thresh]{autothresholdr::mean_stack_thresh()}}) to use on the image prior to
detrending and brightness calculations.}

\item{fail}{If thresholding is done, to which value should pixels not
exceeding the threshold be set?}

\item{filt}{Do you want to smooth (\code{filt = 'mean'}) or median (\code{filt = 'median'}) filter the number image using \code{\link[=smooth_filter]{smooth_filter()}} or
\code{\link[=median_filter]{median_filter()}} respectively? If selected, these are invoked here with a
filter radius of 1 (with corners included, so each median is the median of
9 elements) and with the option \code{na_count = TRUE}. If you want to
smooth/median filter the number image in a different way, first calculate
the numbers without filtering (\code{filt = NULL}) using this function and then
perform your desired filtering routine on the result.}

\item{s}{A number. The \eqn{S}-factor of microscope acquisition.}

\item{offset}{Microscope acquisition parameters. See reference
Dalal et al.}

\item{readout_noise}{Microscope acquisition parameters. See reference
Dalal et al.}

\item{gamma}{Factor for correction of number \eqn{n} due to the illumination
profile. The default (\code{gamma = 1}) has no effect. Changing gamma will have
the effect of dividing the result by \code{gamma}, so the result with \code{gamma = 0.5} is two times the result with \code{gamma = 1}. For a Gaussian illumination
profile, use \code{gamma = 0.3536}; for a Gaussian-Lorentzian illumination
profile, use \code{gamma = 0.0760}.}

\item{seed}{Random numbers may be generated during the detrending process.
For reproducibility, you can set a seed for this random number generation
here.}

\item{parallel}{Would you like to use multiple cores to speed up this
function? If so, set the number of cores here, or to use all available
cores, use \code{parallel = TRUE}.}
}
\value{
A matrix, the brightness image.
}
\description{
Given a time stack of images, \code{brightness()} performs a calculation of the
brightness for each pixel. \code{brightness_txt_folder()} does this calculation
for an entire folder, writing the results as text files via
\code{\link[=write_txt_img]{write_txt_img()}}.
}
\examples{
img <- read_tif(system.file('extdata', '50.tif', package = 'nandb'))
display(img[, , 1])
num <- brightness(img, "e", tau = NA, thresh = "Huang")
num <- brightness(img, "B", tau = 10, thresh = "tri")
}
\references{
Digman MA, Dalal R, Horwitz AF, Gratton E. Mapping the Number of
Molecules and Brightness in the Laser Scanning Microscope. Biophysical
Journal. 2008;94(6):2320-2332. \doi{10.1529/biophysj.107.114645}.

Dalal, RB, Digman, MA, Horwitz, AF, Vetri, V, Gratton, E (2008).
Determination of particle number and brightness using a laser scanning
confocal microscope operating in the analog mode. Microsc. Res. Tech., 71,
1:69-81. \doi{10.1002/jemt.20526}.
}
