% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brightness.R
\name{BrightnessTimeSeries}
\alias{BrightnessTimeSeries}
\title{Create a Brightness time-series.}
\usage{
BrightnessTimeSeries(mat3d, frames.per.set, pbt = NULL, tau = NA)
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series. To perform this on a file
that has not yet been read in, set this argument to the path to that file
(a string).}

\item{frames.per.set}{The number of frames with which to calculate the
successive brightnesses.}

\item{tau}{The time constant for the exponential filtering.}
}
\value{
An array where the \eqn{i}th slice is the \eqn{i}th brightness image.
}
\description{
Given a stack of images, use the first \code{frames.per.set} of them to
create one brightness image, the next \code{frames.per.set} of them to create
the next brightness image and so on to get a time-series of brightness
images. If \code{tau} is specified, bleaching correction is performed via
\code{\link{CorrectForBleaching}}.
}
\details{
This may discard some images, for example if 175 frames are in the input and
\code{frames.per.set = 50}, then the last 25 are discarded. If bleaching
correction is selected, it is performed on the whole image stack before the
sectioning is done for calculation of brightnesses.
}
\seealso{
\code{\link{Brightness}}.
}

