% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{CorrectForBleaching}
\alias{CorrectForBleaching}
\title{Detrend an image series}
\usage{
CorrectForBleaching(mat3d, tau)
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series. To perform this on a file
that has not yet been read in, set this argument to the path to that file
(a string).}

\item{tau}{If this is specified, bleaching correction is performed with
]code{\link{CorrectForBleaching}} which uses exponential filtering with
time constant \code{tau} (where the unit of time is the time between
frames).}
}
\value{
The detrended image series.
}
\description{
Apply detrending to an image time series using the bleaching correction
method described in Digman et al. 2008.
}

