% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intensity.R
\name{MeanIntensity}
\alias{MeanIntensity}
\title{Calculate mean intensity from image series.}
\usage{
MeanIntensity(mat3d, mst = NULL, filt = NULL, verbose = TRUE)
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series. To perform this on a file
that has not yet been read in, set this argument to the path to that file
(a string).}

\item{mst}{Do you want to apply an intensity threshold prior to calculating
mean intensities (via \code{\link{MedStackThresh}})? If so, set your
thresholding \emph{method} here.}

\item{filt}{Do you want to median/smooth filter (with a radius of 1) the
resulting image using \code{\link{MedianFilterB}} or
\code{\link{SmoothFilterB}}?}

\item{verbose}{If mat3d is specified as a file name, print a message to tell
the user that that file is now being processed (useful for
\code{MeanIntensityTxtFolder}, does not work with multiple cores) and to
tell when \code{MeanIntensityTxtFolder} is done.}
}
\value{
\code{Brightness} returns a matrix, the brightness image. The result
  of \code{BrightnessTxtFolder} is the text csv files written to disk (in the
  same folder as the input images).
}
\description{
Given a time stack of images, \code{MeanIntensity} calculates the mean
intensity for each pixel, returning a matrix. \code{MeanIntensityTxtFolder}
does this for every image in a folder, writing the results as text files via
\code{\link{WriteImageTxt}}.
}
\examples{
library(EBImage)
img <- ReadImageData(system.file("extdata",
"low_oligomers.tif",
package = "nandb"))
display(normalize(img[, , 1]), method = "raster")
mean.intensity <- MeanIntensity(img, mst = "Huang", filt = "median")
display(normalize(mean.intensity), method = "r")
}

