% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{MedStackThresh}
\alias{MedStackThresh}
\title{Threshold every image in a stack based on their median.}
\usage{
MedStackThresh(mat3d, method = "Huang")
}
\arguments{
\item{mat3d}{A 3-dimensional array (the image stack) where the \eqn{n}th
slice is the \eqn{n}th image in the time series.}

\item{method}{The thresholding method to use. See
\code{\link[autothresholdr]{auto_thresh_mask}}.}
}
\value{
The thresholded stack, pillars not exceeding the threshold are set to
  zero.
}
\description{
Via the \code{\link[autothresholdr]{auto_thresh_mask}} function from the
\code{autothresholdr} package, apply a threshold to the image stack before
calculating brightness, such as to only have nonzero brightnesses in areas
where the cell is.
}
\details{
This function takes the median of all the slices in the stack and uses this
image to create a mask via \code{\link[autothresholdr]{auto_thresh_mask}} and
then applies this mask to every slice in the stack (so for a given pillar in
the image stack, either it is wiped away to zero, or it is untouched).
}
\examples{
library(EBImage)
img <- ReadImageData(system.file("extdata",
"FKBP-mClover_before_0.5nM_AP1510.tif",
package = "nandb"))
display(normalize(img[, , 1]), method = "raster")
img_thresh_mask <- MedStackThresh(img, "Otsu")
display(img_thresh_mask[, , 1] > 0, method = "r")
display(normalize(img[, , 1]), method = "raster")
img_thresh_mask <- MedStackThresh(img, "Triangle")
display(img_thresh_mask[, , 1] > 0, method = "r")

}

