% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadImageData}
\alias{ReadImageData}
\title{Read image as array object.}
\usage{
ReadImageData(image.name, restore.counts = TRUE)
}
\arguments{
\item{image.name}{The path to the image file on disk. The file extension must
be one of ".jpeg", ".png", ".tiff" or ".tif".}

\item{restore.counts}{If this is \code{TRUE} (the default), pixel intensities
are read in as integers in the range [0, 2 ^ bits - 1]. If set to
\code{FALSE}, the default behaviour of \code{\link[EBImage]{readImage}} is
restored, however the resulting array still has a "bits" attribute, which
it wouldn't with \code{\link[EBImage]{readImage}}.}
}
\value{
An array with a "bits" attribute.
}
\description{
Read in an image file from the disk as an array of pixel intensities. Give it
an attribute "bits" detailing how many bits per sample there were in the file
on disk. optionally allows the user to read in pixel intensities as integers
in the range [0, 2 ^ bits - 1]. See "Details".
}
\details{
This function wraps \code{\link[EBImage]{readImage}} and
\code{\link[EBImage]{imageData}} from the \code{\link[EBImage]{EBImage}}
package. By default, \code{readImage} reads in pixel intensities in the range
[0, 1]. \code{ReadImageData} optionally allows the user to read in pixel
intensities as integers in the range [0, 2 ^ bits - 1], as would be seen when
viewing the file in an application such as imagej. This is sometimes
necessary, for example, when calculating number and brightness, where we need
pixel values to be in units of "counts". This functionality is enabled by
default, but can be disabled with \code{restore.counts = FALSE}.

This function deals with reading in image stacks in the same way as
\code{\link[EBImage]{EBImage}}.
}

